Added delete lesson

+++ modules/AmirVahedix/Course/Repositories/CourseRepo.php
@@ -29,7 +29,7 @@ class CourseRepo
     public function update(Course $course, UpdateCourseRequest $request)
     {
         if ($request->has('banner')) {
-            $course->banner->delete();
+            if ($course->banner) $course->banner->delete();
             $banner_id = MediaService::publicUpload($request->file('banner'))->id;
             $request->request->add(['banner_id' => $banner_id]);
         } else {
